requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-groq>=0.2.0
bleach==6.1.0
Flask-Limiter==3.7.0
Flask-Caching==2.3.0
flask-compress==1.14
google-cloud-language==2.13.0
google-cloud-translate==3.15.3
google-cloud-vision==3.7.0
google-cloud-bigquery==3.25.0
firebase-admin==6.5.0
pytest==8.2.2
pytest-cov==5.0.0
